requirements.txt
discord.py==1.7.2
discord.py[voice]==1.7.2
dpymenus==2.1.3
firebase-admin==4.5.3
Pillow==8.2.0
